When generating COUNT, don't add a DISTINCT unless we need it
A simple check of ->distinct_required is not sufficient -- when
building joins for prefetch, we explicitly mark them as "distinct"
joins, because we want all of their table's rows back for every one of
the primary row.  When doing a count, however, we either need to
un-join the prefetch tables, or simply elect to add a DISTINCT.

Since this is still strictly better than previously, in the common
case of actual distinct joins, this corner case is acceptable, and
indeed no worse than it was previously.