Add an attribute which controls placeholder use for load_by_cols
Postgres allows for partial indexes, but can only use them if the
values are known at prepare time, when the query plan is generated.
The no_placeholder attribute tells load_by_cols to insert the value
into the statement directly (quoted, of course), instead of at
execute-time, using a placeholder.  This allows Postgres to make use
of partial indexes on the given column.