Set raw_values when we set values and fetched => 1
Otherwise raw_values is always empty when load_by_hash is used.  The
most common use case for load_by_hash is instantiating records in a
collection object.  Note that decoded is still false for all columns.

Before any columns are decoded, values and raw_values should be the same
(as other places in the code ensure).