Avoid a compilation failure caused by incorrect precedence
The intended assignment on the else side of the ternary operator needs
to be in quotes since ?: binds tighter than =.  Rather than just add the
parens and be done with it, clarify the code a little.

Apparently Perl previously passed this by without making a fuss.  It
started failing to compile on blead as of v5.15.6-118-g2846acb.  See the
bug report at rt.cpan.org #74298.