Fix for @{ eval { } } parsing under perl 5.19.4 and above
Under 5.19.4, @{ foo {"bar"} } is always parsed as @foo{bar} -- whereas
in previous version of perl, it was parsed as either that or
@{ ( foo{"bar"} ) }, depending on whenever there was a newline after the
'@{', and on whenever you were inside of an eval STRING or not.

Jifty::DBI::Collection used '@{\neval {} ...}', which under perl 5.19.4
and above parses as a syntax error.  As such, explicitly mark the inside
of @{} as an expression using parens.